rust: iopoll: add `read_poll_timeout` function
This patch adds `read_poll_timeout()` macro rewritten in Rust.
It is intended for indirect use by wrapping any specialization of
this macro in a concrete function definition such as `readx_poll_timeout()`
for a given `read()` function. This provides additional type-safety for the interface
as well as encapsulates the `read_poll_timeout!()` macro.

Signed-off-by: Maciej Falkowski <m.falkowski@samsung.com>